Separated win2008 and win2012 function wrappers. Dhcpsapi gem will now always load, with api level reflected in DhcpsApi::Server::API_LEVEL constant.
