+++ integration_test/test_helper.rb
@@ -1,3 +1,4 @@
+require "minitest/autorun"
 require "test/unit"
 require 'fileutils'
