+++ lib/dhcpsapi.rb
@@ -1,16 +1,14 @@
 require 'ffi'
-require 'dhcpsapi/ffi'
 require 'dhcpsapi/enums'
 require 'dhcpsapi/version'
 require 'dhcpsapi/common'
-require 'dhcpsapi/common_data_structs'
+require 'dhcpsapi/data_structures'
 require 'dhcpsapi/error'
 require 'dhcpsapi/class'
-require 'dhcpsapi/option'
-require 'dhcpsapi/option_value'
-require 'dhcpsapi/subnet'
-require 'dhcpsapi/subnet_element'
-require 'dhcpsapi/reservation'
 require 'dhcpsapi/client'
 require 'dhcpsapi/misc'
+require 'dhcpsapi/option'
+require 'dhcpsapi/option_value'
+require 'dhcpsapi/reservation'
+require 'dhcpsapi/subnet'
 require 'dhcpsapi/server'
