--- lib/dhcpsapi/ffi.rb
@@ -1,6 +1,0 @@
-module DhcpsApi
-  extend FFI::Library
-
-  ffi_lib 'dhcpsapi'
-  ffi_convention :stdcall
-end
