+++ lib/dhcpsapi/misc.rb
@@ -1,16 +1,4 @@
 module DhcpsApi
-=begin
-  DWORD DHCP_API_FUNCTION DhcpV4GetFreeIPAddress(
-    _In_opt_ LPWSTR          ServerIpAddress,
-    _In_     DHCP_IP_ADDRESS ScopeId,
-    _In_     DHCP_IP_ADDRESS startIP,
-    _In_     DHCP_IP_ADDRESS endIP,
-    _In_     DWORD           numFreeAddrReq,
-    _Out_    LPDHCP_IP_ARRAY *IPAddrList
-  );
-=end
-  attach_function :DhcpV4GetFreeIPAddress, [:pointer, :uint32, :uint32, :uint32, :uint32, :pointer], :uint32
-
   module Misc
     # Returns free ip addresses as a List of Strings.
     #
@@ -27,12 +15,12 @@ module DhcpsApi
     #
     def get_free_ip_address(subnet_address, start_address = nil, end_address = nil, num_of_addresses = 1)
       dhcp_ip_array_ptr_ptr = FFI::MemoryPointer.new(:pointer)
-      error = DhcpsApi.DhcpV4GetFreeIPAddress(to_wchar_string(server_ip_address),
-                                      ip_to_uint32(subnet_address),
-                                      start_address.nil? ? 0 : ip_to_uint32(start_address),
-                                      end_address.nil? ? 0 : ip_to_uint32(end_address),
-                                      num_of_addresses,
-                                      dhcp_ip_array_ptr_ptr)
+      error = DhcpsApi::Win2012::Misc.DhcpV4GetFreeIPAddress(to_wchar_string(server_ip_address),
+                                              ip_to_uint32(subnet_address),
+                                              start_address.nil? ? 0 : ip_to_uint32(start_address),
+                                              end_address.nil? ? 0 : ip_to_uint32(end_address),
+                                              num_of_addresses,
+                                              dhcp_ip_array_ptr_ptr)
 
       return [] if (error == 2 || error == 20126)
       if is_error?(error)
