+++ lib/dhcpsapi/version.rb
@@ -1,3 +1,3 @@
 module DhcpsApi
-  VERSION = "0.0.3"
+  VERSION = "0.0.4"
 end
