--- lib/dhcpsapi_for_testing.rb
@@ -1,3 +1,0 @@
-require 'dhcpsapi/enums'
-require 'dhcpsapi/error'
-require 'dhcpsapi_for_testing/server'
